refactor, getting ready for release
