+++ README.md
@@ -4,6 +4,41 @@ JavaScript object that creates a unique CSS selector for a given DOM element.
 
 It also generates shorter selectors and is faster and/or more robust than many other libraries - see this [comparison](https://github.com/fczbkk/css-selector-generator-benchmark) and select the best alternative for your use case.
 
+## Install
+
+Add the library to your project via NPM...
+
+```shell
+npm install css-selector-generator
+```
+
+...or Yarn.
+
+```shell
+yarn add css-selector-generator
+```
+
+Then include it in your source code:
+
+```javascript
+import getCssSelector from 'css-selector-generator'
+``` 
+
+## How to use
+
+Simplest way to use it is to provide an element reference, without any options.
+
+```html
+<body>
+	<div class="myElement"></div>
+</body>
+```
+
+```javascript
+getCssSelector(document.body.firstElementChild)
+// ".myElement"
+```
+
 ## Documentation
 
 <!-- Generated by documentation.js. Update this documentation by updating the source code. -->
@@ -34,11 +69,12 @@ Type: [Object](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Globa
 #### Properties
 
 -   `selectors` **[Array](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array)&lt;[css_selector_type](#css_selector_type)>?** List of selector types to use. They will be prioritised by their order.
--   `whitelist` **([RegExp](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/RegExp) \| [string](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String))?** List of selectors that should be prioritised.
--   `blacklist` **([RegExp](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/RegExp) \| [string](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String))?** List of selectors that should be ignored.
+-   `whitelist` **[Array](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array)&lt;([RegExp](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/RegExp) \| [string](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String))>?** List of selectors that should be prioritised.
+-   `blacklist` **[Array](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array)&lt;([RegExp](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/RegExp) \| [string](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String))>?** List of selectors that should be ignored.
 -   `root` **[Element](https://developer.mozilla.org/docs/Web/API/Element)?** Root element inside which the selector will be generated. If not set, the document root will be used.
 -   `combineWithinSelector` **[boolean](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Boolean)?** If set to `false`, the generator will not test combinations of selectors of single type (e.g. multiple class selectors).
 -   `combineBetweenSelectors` **[boolean](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Boolean)?** If set to `true`, the generator will try to test combinations of selectors of different types. Warning: Turning this on may cause performance issues.
+-   `includeTag` **[boolean](https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Boolean)?** If set to `true`, all generated selectors will include the TAG part. Even if tag selector type is not included in `selectors` option.
 
 ### css_selector_type
