+++ test/complex.spec.js
@@ -1,6 +1,6 @@
 import {assert} from 'chai';
 import {getCssSelector} from './../src/index';
-import {testSelector} from '../src/utilities';
+import {testSelector} from '../src/utilities-dom';
 import html_code from './html/complex.html';
 
 describe('complex test', function () {
